+++ SBVUnitTest/SBVUnitTestBuildTime.hs
@@ -2,4 +2,4 @@
 module SBVUnitTestBuildTime (buildTime) where
 
 buildTime :: String
-buildTime = "Tue May  5 19:08:05 PDT 2015"
+buildTime = "Tue May  5 21:47:28 PDT 2015"